This commit: Fixed a unique bug where my api-key had become the store name for a store in avault. Learning error.
Removed all instances of javascript 6 type code from my implimentation.
Lots of default arguments gone. Also, not using a promise but instead using
a callback function to deliver data downstream.

This latest commit is up on apigee at http://chutch-test.apigee.net/github_pa3
and it is secured by oauth 2.0 at this time. Unless I need to reupload code due to fixing
an error in a test, you should not be able to access it.